+++ src/front/apps/fl-find/index.js
@@ -1,6 +1,6 @@
 import React, { Fragment, useState } from 'react'
 import { PostList } from 'components/post-list'
-import { TagGroup, Tag, ScreenHeader, ExpandedContents, ContentList } from 'components'
+import { TagGroup, Tag, ScreenHeader, ExpandedContents } from 'components'
 
 export const FindTab = props => {
 
@@ -52,8 +52,7 @@ export const FindTab = props => {
                 </ExpandedContents>
 
             </ScreenHeader>
-			<ContentList
-				type="posts"
+			<PostList
 				query={ {
 					post_type: 'page',
 					numberposts: -1,
