+++ src/front/ui/components/content-list/style.scss
@@ -1,0 +1,9 @@
+.fl-asst-list {
+    list-style: none;
+    padding:0;
+}
+
+.fl-asst-list-loading {
+	padding: 40px;
+	text-align: center;
+}
